Add one docs template for now, use gatsby node api for url pathing
Changed the way url pathing works a bit. Supports making the path
frontmatter field optional. If it's omitted, in gatsby config we check
if its omitted. we set it in the path field in createPage otherwise
gatsby gets mad. and we add a custom node field called urlPath, mirror
that to what we do in createPage, and we query that type in graphql in
the BasicDocPageTemplate.